requirements.txt
google-genai==1.53.0
pypdf==6.4.0
pdfplumber==0.11.8
pandas==2.3.3
openpyxl==3.1.5
pydantic==2.12.5
